+++ src/registry/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
-import { BaseControllerMethods, Controller, ControllerMethods, ControllerProps, CRUDBase } from '../';
+import { BaseControllerMethods, Controller, ControllerMethods, ControllerProps, CRUDBase, SubController } from '../';
 
 type PartialParameter<T extends (...args: unknown[]) => unknown> = Partial<Parameters<T>[0]> extends infer R
 	? R extends undefined
@@ -9,6 +9,10 @@ type PartialParameter<T extends (...args: unknown[]) => unknown> = Partial<Param
 type PartialFunction<T extends (...args: unknown[]) => unknown> = PartialParameter<T> extends infer R
 	? (arg0: R) => ReturnType<T>
 	: never;
+
+type PickByType<T, Value> = {
+	[P in keyof T as T[P] extends Value | undefined ? P : never]: T[P];
+};
 
 /**
  * Allows to register every controller database handler
@@ -26,32 +30,37 @@ export const controllerRegistry = () => {
 	 * @returns
 	 */
 	const register = <CONT extends Controller<unknown, CRUDBase, CRUDBase>>(
-		c: CONT,
+		controller: CONT,
 		db: {
 			[method in keyof ControllerMethods<unknown, CRUDBase, CRUDBase>]?: PartialFunction<CONT[method]>;
 		} & {
-			[subs in keyof Omit<CONT, keyof ControllerMethods<unknown, CRUDBase, CRUDBase> | keyof ControllerProps>]: {
-				// @ts-ignore
-				[k in keyof CONT[subs] & keyof BaseControllerMethods<unknown>]: PartialFunction<CONT[subs][k]>;
+			[sk in keyof PickByType<
+				Omit<CONT, keyof ControllerMethods<unknown, CRUDBase, CRUDBase> | keyof ControllerProps>,
+				ReturnType<SubController<unknown>>
+			>]: {
+				[k in keyof CONT[sk] & keyof BaseControllerMethods<unknown>]: CONT[sk][k];
 			};
 		}
 	) => {
 		const { index, create, destroy, read, update, ...rest } = db;
+		// Create copy of controller so we won't delete original keys later
+		const copy = { ...controller };
 
-		const subs = Object.keys(rest).reduce<{ [k: string]: BaseControllerMethods<unknown> }>((p, s) => {
-			p[c[s].$url] = rest[s];
+		Object.keys(rest).forEach(k => {
+			// Remap properties
+			copy[k] = { ...copy[k], ...rest[k] };
 
-			return p;
-		}, {} as never);
+			// Remap key name
+			delete Object.assign(copy, { [copy[k].$url]: copy[k] })[k];
+		});
 
-		routes.set(c.$url, {
-			...c,
+		routes.set(copy.$url, {
+			...copy,
 			index,
 			create,
 			destroy,
 			read,
 			update,
-			...subs,
 		});
 
 		return { register, handle };